feat: 이벤트 설명 컴포넌트
